feat: #58 장작 삭제, Image API 모두 Log API 로 통일
