+++ src/main/java/com/macrosoft/modakserver/domain/log/controller/LogController.java
@@ -9,12 +9,14 @@ import com.macrosoft.modakserver.global.BaseResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -26,6 +28,7 @@ import org.springframework.web.bind.annotation.RestController;
 @Tag(name = "Log API", description = "모닥불의 장작 관련 API 입니다.")
 public class LogController {
     public static final String API_CAMPFIRES_LOG = "/{campfirePin}/logs";
+    public static final String API_CAMPFIRES_LOG_IMAGES = "/{campfirePin}/images";
     private final LogService logService;
 
     @Operation(summary = "모닥불의 장작들의 메타데이터 정보 가져오기", description = "모닥불에 업로드 되어 있는 장작들의 메타데이터 (위치, 시간) 정보를 가져옵니다. `SelectMergeLogs` 화면에서 추천 장작을 선정할 때 사용됩니다.")
@@ -86,5 +89,59 @@ public class LogController {
             @PathVariable("logId") long logId
     ) {
         return BaseResponse.onSuccess(logService.removeLog(userDetails.getMember(), campfirePin, logId));
+    }
+
+    @Operation(summary = "사진 상세보기", description = "모닥불에 업로드된 사진의 세부 정보를 봅니다. 사진의 메타데이터와 감정표현을 불러옵니다. 사진을 눌러서 크게 본 화면에서 사용됩니다.")
+    @GetMapping(API_CAMPFIRES_LOG_IMAGES + "/{imageId}")
+    public BaseResponse<LogResponse.ImageDetail> getImageDetail(
+            @AuthenticationPrincipal CustomUserDetails userDetails,
+            @Parameter(description = "모닥불 핀 번호", example = "111111")
+            @PathVariable("campfirePin") int campfirePin,
+            @Parameter(description = "사진 아이디", example = "1")
+            @PathVariable("imageId") Long imageId
+    ) {
+        return BaseResponse.onSuccess(logService.getImageDetail(userDetails.getMember(), campfirePin, imageId));
+    }
+
+    @Operation(summary = "감정 표현 하기", description = "사진에 대한 감정 표현을 등록합니다.")
+    @PutMapping(API_CAMPFIRES_LOG_IMAGES + "/{imageId}/emotions")
+    public BaseResponse<LogResponse.ImageDTO> emotion(
+            @AuthenticationPrincipal CustomUserDetails userDetails,
+            @Parameter(description = "모닥불 핀 번호", example = "111111")
+            @PathVariable("campfirePin") int campfirePin,
+            @Parameter(description = "사진 아이디", example = "1")
+            @PathVariable("imageId") Long imageId,
+            @RequestBody LogRequest.EmotionDTO emotionDTO
+    ) {
+        return BaseResponse.onSuccess(
+                logService.emotion(userDetails.getMember(), campfirePin, imageId, emotionDTO.emotion()));
+    }
+
+    @Operation(summary = "감정 표현 삭제", description = "사진에 대한 사용자의 감정 표현을 삭제합니다.")
+    @DeleteMapping(API_CAMPFIRES_LOG_IMAGES + "/{imageId}/emotions")
+    public BaseResponse<LogResponse.ImageDTO> deleteEmotion(
+            @AuthenticationPrincipal CustomUserDetails userDetails,
+            @Parameter(description = "모닥불 핀 번호", example = "111111")
+            @PathVariable("campfirePin") int campfirePin,
+            @Parameter(description = "사진 아이디", example = "1")
+            @PathVariable("imageId") Long imageId
+    ) {
+        return BaseResponse.onSuccess(
+                logService.deleteEmotion(userDetails.getMember(), campfirePin, imageId)
+        );
+    }
+
+    @Operation(summary = "장작에서 사진 삭제하기", description = "장작에서 사진들을 삭제합니다.")
+    @DeleteMapping(API_CAMPFIRES_LOG + "/{logId}/images")
+    public BaseResponse<LogResponse.ImageIds> removeImages(
+            @AuthenticationPrincipal CustomUserDetails userDetails,
+            @Parameter(description = "모닥불 핀 번호", example = "111111")
+            @PathVariable("campfirePin") int campfirePin,
+            @Parameter(description = "장작 아이디", example = "1")
+            @PathVariable("logId") long logId,
+            @RequestParam(value = "imageIds") List<Long> imageIds
+    ) {
+        return BaseResponse.onSuccess(
+                logService.removeImages(userDetails.getMember(), campfirePin, logId, imageIds));
     }
 }
