+++ src/main/java/com/macrosoft/modakserver/domain/log/entity/Emotion.java
@@ -1,4 +1,4 @@
-package com.macrosoft.modakserver.domain.image.entity;
+package com.macrosoft.modakserver.domain.log.entity;
 
 import com.macrosoft.modakserver.domain.member.entity.Member;
 import com.macrosoft.modakserver.global.BaseEntity;
