+++ src/main/java/com/macrosoft/modakserver/domain/log/entity/Log.java
@@ -1,7 +1,6 @@
 package com.macrosoft.modakserver.domain.log.entity;
 
 import com.macrosoft.modakserver.domain.campfire.entity.Campfire;
-import com.macrosoft.modakserver.domain.image.entity.LogImage;
 import com.macrosoft.modakserver.domain.log.dto.LogRequest.ImageInfo;
 import com.macrosoft.modakserver.domain.log.dto.LogRequest.UploadLog;
 import com.macrosoft.modakserver.domain.log.dto.LogResponse.LogMetadata;
