+++ src/main/java/com/macrosoft/modakserver/domain/log/repository/LogImageRepository.java
@@ -1,7 +1,7 @@
-package com.macrosoft.modakserver.domain.image.repository;
+package com.macrosoft.modakserver.domain.log.repository;
 
-import com.macrosoft.modakserver.domain.image.entity.LogImage;
 import com.macrosoft.modakserver.domain.log.entity.Log;
+import com.macrosoft.modakserver.domain.log.entity.LogImage;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
