+++ src/main/java/com/macrosoft/modakserver/domain/log/service/LogService.java
@@ -4,7 +4,9 @@ import com.macrosoft.modakserver.domain.log.dto.LogRequest;
 import com.macrosoft.modakserver.domain.log.dto.LogResponse;
 import com.macrosoft.modakserver.domain.log.dto.LogResponse.LogMetadataList;
 import com.macrosoft.modakserver.domain.log.entity.Log;
+import com.macrosoft.modakserver.domain.log.entity.LogImage;
 import com.macrosoft.modakserver.domain.member.entity.Member;
+import java.util.List;
 
 public interface LogService {
     LogMetadataList getLogsMetadata(Member member, int campfirePin);
@@ -16,6 +18,16 @@ public interface LogService {
     LogResponse.LogDetails getLogDetails(Member member, int campfirePin, Long logId, int page, int size);
 
     LogResponse.LogId removeLog(Member member, int campfirePin, Long logId);
+
+    LogResponse.ImageDetail getImageDetail(Member member, int campfirePin, Long imageId);
+
+    LogResponse.ImageDTO emotion(Member member, int campfirePin, Long imageId, String emotion);
+
+    LogResponse.ImageDTO deleteEmotion(Member member, int campfirePin, Long imageId);
+
+    LogResponse.ImageIds removeImages(Member member, int campfirePin, Long logId, List<Long> imageIds);
+
+    LogImage getLogImage(Long imageId);
 
     Log getLog(Long logId);
 }
