+++ src/main/java/com/macrosoft/modakserver/domain/log/service/LogServiceImpl.java
@@ -1,29 +1,39 @@
 package com.macrosoft.modakserver.domain.log.service;
 
+import static com.macrosoft.modakserver.domain.log.exception.LogErrorCode.EMOTE_EMPTY;
+import static com.macrosoft.modakserver.domain.log.exception.LogErrorCode.EMOTE_TOO_LONG;
+import static com.macrosoft.modakserver.domain.log.exception.LogErrorCode.EMOTION_NOT_FOUND;
+import static com.macrosoft.modakserver.domain.log.exception.LogErrorCode.EMOTION_NOT_UPLOAD_USER;
 import static com.macrosoft.modakserver.domain.log.exception.LogErrorCode.LOG_CAMPFIRE_NOT_MATCH;
+import static com.macrosoft.modakserver.domain.log.exception.LogErrorCode.LOG_IMAGE_NOT_FOUND;
+import static com.macrosoft.modakserver.domain.log.exception.LogErrorCode.LOG_IMAGE_NOT_IN_CAMPFIRE;
 import static com.macrosoft.modakserver.domain.log.exception.LogErrorCode.LOG_NOT_FOUND;
 
 import com.macrosoft.modakserver.domain.campfire.entity.Campfire;
 import com.macrosoft.modakserver.domain.campfire.repository.CampfireRepository;
 import com.macrosoft.modakserver.domain.campfire.service.CampfireService;
-import com.macrosoft.modakserver.domain.image.dto.ImageResponse;
-import com.macrosoft.modakserver.domain.image.entity.LogImage;
-import com.macrosoft.modakserver.domain.image.repository.EmotionRepository;
-import com.macrosoft.modakserver.domain.image.repository.LogImageRepository;
+import com.macrosoft.modakserver.domain.image.service.ImageService;
 import com.macrosoft.modakserver.domain.log.dto.LogRequest;
 import com.macrosoft.modakserver.domain.log.dto.LogResponse;
+import com.macrosoft.modakserver.domain.log.dto.LogResponse.ImageDTO;
+import com.macrosoft.modakserver.domain.log.dto.LogResponse.ImageIds;
 import com.macrosoft.modakserver.domain.log.dto.LogResponse.LogId;
 import com.macrosoft.modakserver.domain.log.dto.LogResponse.LogOverview;
+import com.macrosoft.modakserver.domain.log.entity.Emotion;
 import com.macrosoft.modakserver.domain.log.entity.Location;
 import com.macrosoft.modakserver.domain.log.entity.Log;
-import com.macrosoft.modakserver.domain.log.repository.LocationRepository;
+import com.macrosoft.modakserver.domain.log.entity.LogImage;
+import com.macrosoft.modakserver.domain.log.repository.EmotionRepository;
+import com.macrosoft.modakserver.domain.log.repository.LogImageRepository;
 import com.macrosoft.modakserver.domain.log.repository.LogRepository;
 import com.macrosoft.modakserver.domain.member.entity.Member;
 import com.macrosoft.modakserver.domain.member.service.MemberService;
 import com.macrosoft.modakserver.global.exception.CustomException;
+import jakarta.persistence.EntityManager;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -38,13 +48,15 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class LogServiceImpl implements LogService {
+    public static final int MAX_EMOTE_LENGTH = 10;
     private final CampfireService campfireService;
     private final MemberService memberService;
-    private final EmotionRepository emotionRepository;
-    private final LocationRepository locationRepository;
     private final LogImageRepository logImageRepository;
     private final CampfireRepository campfireRepository;
     private final LogRepository logRepository;
+    private final ImageService imageService;
+    private final EmotionRepository emotionRepository;
+    private final EntityManager entityManager;
 
     @Override
     @Transactional
@@ -227,13 +239,14 @@ public class LogServiceImpl implements LogService {
         return new LogResponse.LogDetails(
                 log.getId(),
                 logImages.stream()
-                        .map(logImage -> new ImageResponse.ImageName(logImage.getId(), logImage.getName()))
+                        .map(logImage -> new LogResponse.ImageName(logImage.getId(), logImage.getName()))
                         .toList(),
                 logImages.hasNext()
         );
     }
 
     @Override
+    @Transactional
     public LogId removeLog(Member member, int campfirePin, Long logId) {
         Member memberInDB = memberService.getMemberInDB(member);
         Campfire campfire = campfireService.findCampfireByPin(campfirePin);
@@ -241,8 +254,12 @@ public class LogServiceImpl implements LogService {
         Log log = getLog(logId);
         validateLogInCampfire(log, campfire);
 
+        List<Long> logImageIds = log.getLogImages().stream()
+                .map(LogImage::getId)
+                .toList();
+        removeImages(member, campfirePin, logId, logImageIds);
         logRepository.delete(log);
-        return null;
+        return new LogId(logId);
     }
 
     @Override
@@ -254,6 +271,141 @@ public class LogServiceImpl implements LogService {
     private void validateLogInCampfire(Log log, Campfire campfire) {
         if (!log.getCampfire().equals(campfire)) {
             throw new CustomException(LOG_CAMPFIRE_NOT_MATCH);
+        }
+    }
+
+    @Override
+    public LogResponse.ImageDetail getImageDetail(Member member, int campfirePin, Long imageId) {
+        Member memberInDB = memberService.getMemberInDB(member);
+        Campfire campfire = campfireService.findCampfireByPin(campfirePin);
+        campfireService.validateMemberInCampfire(memberInDB, campfire);
+        LogImage logImage = getLogImage(imageId);
+        validateLogImageInCampfire(campfire, logImage);
+        return LogResponse.ImageDetail.of(logImage);
+    }
+
+    @Override
+    @Transactional
+    public ImageDTO emotion(Member member, int campfirePin, Long imageId, String emotion) {
+        validateEmoteLength(emotion);
+        Member memberInDB = memberService.getMemberInDB(member);
+        Campfire campfire = campfireService.findCampfireByPin(campfirePin);
+        campfireService.validateMemberInCampfire(memberInDB, campfire);
+        LogImage logImage = getLogImage(imageId);
+        validateLogImageInCampfire(campfire, logImage);
+
+        Optional<Emotion> optionalExistingEmotion = findOneEmotion(member, logImage);
+
+        if (optionalExistingEmotion.isPresent()) {
+            Emotion existingEmotion = optionalExistingEmotion.get();
+            existingEmotion.updateEmotion(emotion);
+            emotionRepository.save(existingEmotion);
+        } else {
+            Emotion newEmotion = Emotion.of(emotion, logImage, memberInDB);
+            emotionRepository.save(newEmotion);
+            logImage.addEmote(newEmotion);
+        }
+
+        return ImageDTO.of(logImage);
+    }
+
+    @Override
+    @Transactional
+    public ImageDTO deleteEmotion(Member member, int campfirePin, Long imageId) {
+        Member memberInDB = memberService.getMemberInDB(member);
+        Campfire campfire = campfireService.findCampfireByPin(campfirePin);
+        campfireService.validateMemberInCampfire(memberInDB, campfire);
+        LogImage logImage = getLogImage(imageId);
+        validateLogImageInCampfire(campfire, logImage);
+        Set<Emotion> emotion = findEmotion(member, logImage);
+        for (Emotion e : emotion) {
+            validateEmotionUploader(e, memberInDB);
+        }
+        emotionRepository.deleteAll(emotion);
+        entityManager.flush();
+        logImage = getLogImage(imageId);
+        return ImageDTO.of(logImage);
+    }
+
+    private Set<Emotion> findEmotion(Member member, LogImage logImage) {
+        Set<Emotion> emotions = emotionRepository.findAllByMemberAndLogImage(member, logImage);
+        if (emotions.isEmpty()) {
+            throw new CustomException(EMOTION_NOT_FOUND);
+        }
+        return emotions;
+    }
+
+    private Optional<Emotion> findOneEmotion(Member member, LogImage logImage) {
+        return emotionRepository.findByMemberAndLogImage(member, logImage);
+    }
+
+    private void validateEmotionUploader(Emotion emotion, Member member) {
+        if (!emotion.getMember().equals(member)) {
+            throw new CustomException(EMOTION_NOT_UPLOAD_USER);
+        }
+    }
+
+    private void validateEmoteLength(String emote) {
+        if (emote == null || emote.isBlank()) {
+            throw new CustomException(EMOTE_EMPTY);
+        }
+        if (emote.length() > MAX_EMOTE_LENGTH) {
+            throw new CustomException(EMOTE_TOO_LONG);
+        }
+    }
+
+    @Override
+    @Transactional
+    public ImageIds removeImages(Member member, int campfirePin, Long logId, List<Long> imageIds) {
+        List<Long> deletedLogImages = new ArrayList<>();
+        Member memberInDB = memberService.getMemberInDB(member);
+        Campfire campfire = campfireService.findCampfireByPin(campfirePin);
+        campfireService.validateMemberInCampfire(memberInDB, campfire);
+
+        List<LogImage> logImages = getLogImages(imageIds);
+        Log log = getLog(logId);
+        validateLogImagesInCampfire(campfire, logImages);
+
+        for (Long imageId : imageIds) {
+            LogImage logImage = getLogImage(imageId);
+            String imageName = logImage.getName();
+            removeImage(logImage);
+            imageService.deleteImageFromS3(imageName);
+            deletedLogImages.add(imageId);
+        }
+
+        if (log.getLogImages().isEmpty()) {
+            logRepository.delete(log);
+        }
+
+        return new ImageIds(deletedLogImages);
+    }
+
+    private List<LogImage> getLogImages(List<Long> imageIds) {
+        return imageIds.stream()
+                .map(this::getLogImage)
+                .toList();
+    }
+
+    private void validateLogImagesInCampfire(Campfire campfire, List<LogImage> logImages) {
+        for (LogImage logImage : logImages) {
+            validateLogImageInCampfire(campfire, logImage);
+        }
+    }
+
+    private void removeImage(LogImage logImage) {
+        logImage.getLog().removeLogImage(logImage);
+        logImageRepository.delete(logImage);
+    }
+
+    public LogImage getLogImage(Long imageId) {
+        return logImageRepository.findById(imageId)
+                .orElseThrow(() -> new CustomException(LOG_IMAGE_NOT_FOUND));
+    }
+
+    private void validateLogImageInCampfire(Campfire campfire, LogImage logImage) {
+        if (!logImage.getLog().getCampfire().equals(campfire)) {
+            throw new CustomException(LOG_IMAGE_NOT_IN_CAMPFIRE);
         }
     }
 }
